Made some core data to prevent concurrency problems. Ready for the Udacity review!

+++ Virtual Tourist/PhotoAlbumCollectionViewController.swift
@@ -11,92 +11,105 @@ import CoreData
 
 let reuseIdentifier = "PhotoAlbumCell"
 
-class PhotoAlbumCollectionViewController: UICollectionViewController, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout,
-UIPopoverPresentationControllerDelegate {
+class PhotoAlbumCollectionViewController: UICollectionViewController, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout, UIPopoverPresentationControllerDelegate {
     
     // will be set before segue from the mapView
     var currentLocation: TravelLocation!
     var currentPhoto: Photo!
     var headerImage : UIImage!
-  
     
+    @IBOutlet weak var newCollectionButton: UIBarButtonItem!
     @IBOutlet var photoCollectionView: UICollectionView!
     
+    var allowIndividualDelete = false
     
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-    }
-    
-    
+  
     override func viewWillAppear(animated: Bool) {
         super.viewWillAppear(true)
         
         self.navigationController?.navigationBarHidden = false
-        self.navigationController?.toolbarHidden = true
+        self.setAllowNewCollectionTo(false)
         
         // Fetch the results from coredata
         fetchedResultsController.performFetch(nil)
         fetchedResultsController.delegate = self
-        // check for empty images
-        let qos = Int(DISPATCH_QUEUE_PRIORITY_BACKGROUND.value)
-        dispatch_async(dispatch_get_global_queue(qos, 0)) {
-            for photo in self.currentLocation.photos! {
-                if photo.localImage == nil {
-                    FlickrClient.sharedInstance().getPhotoFromUrlFor(self.currentLocation, photo: photo)
-                }
-            }
-        }
-    
+        
         if fetchedResultsController.fetchedObjects!.count > 0 {
-          self.navigationController?.toolbarHidden = false
+        
+            self.handlePhotoDownload()
+  
         } else {
-            FlickrClient.sharedInstance().getImagesForLocation(self.currentLocation) {succes, message, error in
-                if !succes {
-                    self.showImageErrorWith(message)
-                }
-            }
+            
+            self.getNewPhotoCollection(self)
+        
         }
-
-       
     }
     
     
-    @IBAction func getNewPhotoCollection(sender: UIBarButtonItem) {
+    @IBAction func getNewPhotoCollection(sender: AnyObject?) {
         
-       FlickrClient.sharedInstance().deleteImagesForLocation(currentLocation) {succes, message, error in
+        self.setAllowNewCollectionTo(false)
+        
+        FlickrClient.sharedInstance().getImagesFor(self.currentLocation) {succes, message, error in
             if succes {
-                // get a new batch of images
-                 FlickrClient.sharedInstance().getImagesForLocation(self.currentLocation) {succes, message, error in
-                    
-                    let qos = Int(DISPATCH_QUEUE_PRIORITY_HIGH.value)
-                    dispatch_async(dispatch_get_global_queue(qos, 0)) {
-                        for photo in self.currentLocation.photos! {
-                            if photo.localImage == nil {
-                                FlickrClient.sharedInstance().getPhotoFromUrlFor(self.currentLocation, photo: photo)                            }
-                        }
-                    }
-                }
+
+                self.handlePhotoDownload()
+                
             } else {
+                
                 self.showImageErrorWith(message)
+           
             }
+        }
+    }
+
+
+    func handlePhotoDownload() {
+  
+        if self.fetchedResultsController.fetchedObjects!.count > 0 {
+            for photo in self.fetchedResultsController.fetchedObjects as! [Photo] {
+                if !photo.downloaded {
+                    FlickrClient.sharedInstance().getPhotoFromUrlFor(photo)
+                }
+            }
+            
+            self.setAllowNewCollectionTo(true)
+            
+        } else {
+            
+            var alert = UIAlertView(title: "No images found", message: "Sorry, we found no (new) images for this location.", delegate: nil, cancelButtonTitle: "OK")
+            alert.show()
+            
+        }
+    }
+    
+    
+    func setAllowNewCollectionTo(bool: Bool) {
+   
+        if bool {
+            self.newCollectionButton.enabled = true
+            self.navigationController?.toolbarHidden = false
+            self.allowIndividualDelete = true
+            
+        } else {
+            self.newCollectionButton.enabled = false
+            self.navigationController?.toolbarHidden = true
+            self.allowIndividualDelete = false
         }
     }
     
     
     func showImageErrorWith(message: String) {
         // show the error
-        var composeAlert = UIAlertController(title: "Error in retrieving images", message: "Sorry, we encountered an error : \(message)", preferredStyle: UIAlertControllerStyle.Alert)
-        composeAlert.addAction(UIAlertAction(title: "Ok", style: .Default, handler: { (action: UIAlertAction!) in
-            // no action
-        }))
-        self.presentViewController(composeAlert, animated: true, completion: nil)
+        var alert = UIAlertView(title: "Problem retrieving a collection", message: "Sorry, we encountered an error : \(message)", delegate: nil, cancelButtonTitle: "OK")
+        alert.show()
     }
     
     
     // MARK: - Core Data
     
     var sharedContext: NSManagedObjectContext {
+    
         return CoreDataStackManager.sharedInstance().managedObjectContext!
     }
     
@@ -106,13 +119,15 @@ UIPopoverPresentationControllerDelegate {
     lazy var fetchedResultsController: NSFetchedResultsController = {
         
         let fetchRequest = NSFetchRequest(entityName: "Photo")
-        fetchRequest.sortDescriptors = [NSSortDescriptor(key: "creation", ascending: true)] //sort old->new
+        fetchRequest.sortDescriptors = [NSSortDescriptor(key: "localFilePath", ascending: true)] //sort old->new
         fetchRequest.predicate = NSPredicate(format: "location == %@", self.currentLocation); // only for the current location
         
-        let fetchedResultsController = NSFetchedResultsController(fetchRequest: fetchRequest,
+        let fetchedResultsController = NSFetchedResultsController(
+            fetchRequest: fetchRequest,
             managedObjectContext: self.sharedContext,
             sectionNameKeyPath: nil,
-            cacheName: nil)
+            cacheName: nil
+        )
         
         return fetchedResultsController
         
@@ -120,9 +135,10 @@ UIPopoverPresentationControllerDelegate {
     
     
     //MARK: - popoverdelegate
-    func adaptivePresentationStyleForPresentationController(
-        controller: UIPresentationController) -> UIModalPresentationStyle {
-            return .None
+    func adaptivePresentationStyleForPresentationController(controller: UIPresentationController) -> UIModalPresentationStyle {
+ 
+        return .None
+    
     }
     
     
@@ -137,6 +153,7 @@ UIPopoverPresentationControllerDelegate {
 
 
     override func collectionView(collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+
         let sectionInfo = self.fetchedResultsController.sections![section] as! NSFetchedResultsSectionInfo
         
         return sectionInfo.numberOfObjects
@@ -148,7 +165,7 @@ UIPopoverPresentationControllerDelegate {
         let cell = collectionView.dequeueReusableCellWithReuseIdentifier(reuseIdentifier, forIndexPath: indexPath) as! PhotoAlbumUICollectionViewCell
         let photo = fetchedResultsController.objectAtIndexPath(indexPath) as! Photo
         // check if we have an image or have to show a placeholder
-        if photo.localImage != nil {
+        if photo.downloaded {
             cell.locationImage.image = photo.localImage
             cell.loadImageActivityIndicator.stopAnimating()
         } else {
@@ -172,12 +189,14 @@ UIPopoverPresentationControllerDelegate {
         return cell
     }
     
+    
     func showImageDetails (gestureRecognizer : UIGestureRecognizer) {
+ 
         if let indexPath = self.collectionView?.indexPathForCell(gestureRecognizer.view as! PhotoAlbumUICollectionViewCell) {
             if let photo = fetchedResultsController.objectAtIndexPath(indexPath) as? Photo {
                 currentPhoto = photo
                 if photo.localImage != nil {
-                    performSegueWithIdentifier(Constants.SegueIdentifier.showImageDetails, sender: self) // segue to imagedetails
+                    performSegueWithIdentifier(Constants.SegueIdentifier.ShowImageDetails, sender: self) // segue to imagedetails
                 }
             }
         }
@@ -188,14 +207,11 @@ UIPopoverPresentationControllerDelegate {
     override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
        
         switch segue.identifier! {
-            case Constants.SegueIdentifier.showImageDetails:
+            case Constants.SegueIdentifier.ShowImageDetails:
                 let controller = segue.destinationViewController as! ImageDetailsViewController
                 controller.photo = currentPhoto
-            case Constants.SegueIdentifier.showTags:
+            case Constants.SegueIdentifier.ShowTags:
                 let popoverViewController = segue.destinationViewController as! TagCollectionViewController
-                
-                
-                
                 popoverViewController.currentLocation = self.currentLocation
                 popoverViewController.modalPresentationStyle = UIModalPresentationStyle.Popover
                 popoverViewController.popoverPresentationController!.delegate = self
@@ -206,11 +222,12 @@ UIPopoverPresentationControllerDelegate {
 
     
     func removeFromCollection(gestureRecognizer : UIGestureRecognizer){
-
-        if let indexPath = self.collectionView?.indexPathForCell(gestureRecognizer.view as! PhotoAlbumUICollectionViewCell) {
-            let photoForDelete = fetchedResultsController.objectAtIndexPath(indexPath) as! Photo
-            sharedContext.deleteObject(photoForDelete)
-            CoreDataStackManager.sharedInstance().saveContext()
+        if allowIndividualDelete {
+            if let indexPath = self.collectionView?.indexPathForCell(gestureRecognizer.view as! PhotoAlbumUICollectionViewCell) {
+                let photoForDelete = fetchedResultsController.objectAtIndexPath(indexPath) as! Photo
+                sharedContext.deleteObject(photoForDelete)
+                CoreDataStackManager.sharedInstance().saveContext()
+            }
         }
     }
     
@@ -245,7 +262,8 @@ UIPopoverPresentationControllerDelegate {
     
     
     func collectionView(collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
-          var size = (collectionViewLayout as! UICollectionViewFlowLayout).headerReferenceSize
+        
+        var size = (collectionViewLayout as! UICollectionViewFlowLayout).headerReferenceSize
         size.width = self.collectionView!.frame.width
         size.height = self.collectionView!.frame.height/10
         return size
@@ -254,64 +272,51 @@ UIPopoverPresentationControllerDelegate {
 }
 
 
-
+// MARK - extension PhotoAlbumCollectionViewController
 
 extension PhotoAlbumCollectionViewController: NSFetchedResultsControllerDelegate {
     
-    
     func controllerWillChangeContent(controller: NSFetchedResultsController) {
-       // not implemented
+        //
     }
-
-    
-    func controller(controller: NSFetchedResultsController,
-        didChangeSection sectionInfo: NSFetchedResultsSectionInfo,
-        atIndex sectionIndex: Int,
-        forChangeType type: NSFetchedResultsChangeType) {
-            
-            switch type {
-                case .Insert:
-                    println("insert section")
-                case .Delete:
-                    println("delete section")
-                default:
-                    return
-            }
-    }
-
     
     func controller(controller: NSFetchedResultsController, didChangeObject anObject: AnyObject, atIndexPath indexPath: NSIndexPath?, forChangeType type: NSFetchedResultsChangeType, newIndexPath: NSIndexPath?) {
+ 
         let changedPhoto = anObject as! Photo
+        
         switch type {
-            case .Insert:
-                // image added: show it
+            case .Insert:   // image added: show it
+                
                 dispatch_sync(dispatch_get_main_queue()) {
-                   self.collectionView!.insertItemsAtIndexPaths([newIndexPath!])
-                 }
-            case .Delete:
-                // image deleted: remove it
+                    self.collectionView!.insertItemsAtIndexPaths([newIndexPath!])
+                }
+            
+            case .Delete:     // image deleted: remove it
+    
                 self.photoCollectionView.deleteItemsAtIndexPaths([indexPath!])
-            case .Update:
-                // image updated: refresh it
-                if changedPhoto.localImage != nil {
+            
+            case .Update:   // image updated: refresh it
+                
+                if changedPhoto.downloaded {
+                    
                     dispatch_async(dispatch_get_main_queue()) {
                         if let cell = self.photoCollectionView.cellForItemAtIndexPath(indexPath!) as? PhotoAlbumUICollectionViewCell {
                             cell.locationImage.image = changedPhoto.localImage!
                             cell.loadImageActivityIndicator.stopAnimating()
+                            }
+                        }
+                
+                } else {
+                    
+                    dispatch_async(dispatch_get_main_queue()) {
+                        if let cell = self.photoCollectionView.cellForItemAtIndexPath(indexPath!) as? PhotoAlbumUICollectionViewCell {
+                            cell.locationImage.image = UIImage(named: "placeHolderError")
+                            cell.loadImageActivityIndicator.stopAnimating()
                         }
                     }
                 }
-            
-            //case .Move:
-                // not implemented
-                
-            default:
+        default:
                 return
         }
     }
-
-    func controllerDidChangeContent(controller: NSFetchedResultsController) {
-        //  not implemented
-   }
-    
 }
